requirements_demo.txt
python-dotenv==1.0
gradio==4.32